+++ src/ResourceManager/DataFactoryV2/Commands.DataFactoryV2/Runs/GetAzureDataFactoryActivityRunCommand.cs
@@ -17,11 +17,12 @@ using System.Collections.Generic;
 using System.Management.Automation;
 using System.Security.Permissions;
 using Microsoft.Azure.Commands.DataFactoryV2.Models;
+using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 
 namespace Microsoft.Azure.Commands.DataFactoryV2
 {
     [Cmdlet(VerbsCommon.Get, Constants.ActivityRun, DefaultParameterSetName = ParameterSetNames.ByFactoryName),
-        OutputType(typeof(List<PSActivityRun>), typeof(PSActivityRun))]
+        OutputType(typeof(PSActivityRun))]
     public class GetAzureDataFactoryActivityRunCommand : DataFactoryContextBaseCmdlet
     {
         [Parameter(ParameterSetName = ParameterSetNames.ByFactoryObject, Position = 1, Mandatory = true,
@@ -59,6 +60,7 @@ namespace Microsoft.Azure.Commands.DataFactoryV2
         [ValidateNotNullOrEmpty]
         public string Status { get; set; }
 
+        [CmdletParameterBreakingChange("LinkedServiceName", ChangeDescription = Constants.DeprecatingParameter)]
         [Parameter(ParameterSetName = ParameterSetNames.ByFactoryObject, Position = 6, Mandatory = false,
             HelpMessage = Constants.HelpLinkedServiceName)]
         [Parameter(ParameterSetName = ParameterSetNames.ByFactoryName, Position = 7, Mandatory = false,