+++ src/ResourceManager/RecoveryServices.SiteRecovery/Commands.RecoveryServices.SiteRecovery.Test/ScenarioTests/Common/AsrTests.cs
@@ -24,10 +24,11 @@ namespace RecoveryServices.SiteRecovery.Test
 {
     public class AsrCommonTests : AsrTestsBase
     {
+        public XunitTracingInterceptor _logger;
         public AsrCommonTests(
             ITestOutputHelper output)
         {
-            XunitTracingInterceptor.AddToContext(new XunitTracingInterceptor(output));
+            _logger = new XunitTracingInterceptor(output);
             this.vaultSettingsFilePath = System.IO.Path.Combine(
                 System.AppDomain.CurrentDomain.BaseDirectory,
                 "ScenarioTests\\Common\\Common.VaultCredentials");
@@ -44,6 +45,7 @@ namespace RecoveryServices.SiteRecovery.Test
         public void EnumerationTests()
         {
             this.RunPowerShellTest(
+                _logger,
                 Constants.NewModel,
                 "Test-SiteRecoveryEnumerationTests -vaultSettingsFilePath \"" +
                 this.vaultSettingsFilePath +
@@ -57,17 +59,19 @@ namespace RecoveryServices.SiteRecovery.Test
         public void V2AEvent()
         {
             this.RunPowerShellTest(
+                _logger,
              Constants.NewModel,
              "Test-AsrEvent -vaultSettingsFilePath \"" + this.vaultSettingsFilePath + "\"");
         }
 
-        [Fact(Skip = "Recording pending with new APi version")]
+        [Fact]
         [Trait(
             Category.AcceptanceType,
             Category.CheckIn)]
         public void V2AGetJobTest()
         {
             this.RunPowerShellTest(
+                _logger,
              Constants.NewModel,
              "Test-Job -vaultSettingsFilePath \"" + this.vaultSettingsFilePath + "\"");
         }
@@ -79,6 +83,7 @@ namespace RecoveryServices.SiteRecovery.Test
         public void V2AGetNotificationTest()
         {
             this.RunPowerShellTest(
+                _logger,
              Constants.NewModel,
              "Test-NotificationSettings -vaultSettingsFilePath \"" + this.vaultSettingsFilePath + "\"");
         }