+++ src/ResourceManager/RecoveryServices.SiteRecovery/Commands.RecoveryServices.SiteRecovery.Test/ScenarioTests/E2A/AsrE2ATests.cs
@@ -22,10 +22,12 @@ namespace RecoveryServices.SiteRecovery.Test
 {
     public class AsrE2ATests : AsrTestsBase
     {
+        public XunitTracingInterceptor _logger;
         public AsrE2ATests(
             ITestOutputHelper output)
         {
-            XunitTracingInterceptor.AddToContext(new XunitTracingInterceptor(output));
+            _logger = new XunitTracingInterceptor(output);
+            XunitTracingInterceptor.AddToContext(_logger);
             this.vaultSettingsFilePath = System.IO.Path.Combine(
                 System.AppDomain.CurrentDomain.BaseDirectory,
                 "ScenarioTests\\E2A\\E2A.VaultCredentials");
@@ -42,6 +44,7 @@ namespace RecoveryServices.SiteRecovery.Test
         public void FabricTests()
         {
             this.RunPowerShellTest(
+                _logger,
                 Constants.NewModel,
                 "Test-FabricTest -vaultSettingsFilePath \"" +
                 this.vaultSettingsFilePath +
@@ -55,19 +58,21 @@ namespace RecoveryServices.SiteRecovery.Test
         public void TestCreatePolicy()
         {
             this.RunPowerShellTest(
+                _logger,
                 Constants.NewModel,
                 "Test-SiteRecoveryCreatePolicy -vaultSettingsFilePath \"" +
                 this.vaultSettingsFilePath +
                 "\"");
         }
 
-        [Fact(Skip = "Recording pending with new APi version")]
+        [Fact]
         [Trait(
             Category.AcceptanceType,
             Category.CheckIn)]
         public void TestEnableDR()
         {
             this.RunPowerShellTest(
+                _logger,
                 Constants.NewModel,
                 "Test-SiteRecoveryEnableDR -vaultSettingsFilePath \"" +
                 this.vaultSettingsFilePath +
@@ -81,19 +86,21 @@ namespace RecoveryServices.SiteRecovery.Test
         public void TestMapNetwork()
         {
             this.RunPowerShellTest(
+                _logger,
                 Constants.NewModel,
                 "Test-MapNetwork -vaultSettingsFilePath \"" +
                 this.vaultSettingsFilePath +
                 "\"");
         }
 
-        [Fact(Skip = "Recording pending with new APi version")]
+        [Fact]
         [Trait(
             Category.AcceptanceType,
             Category.CheckIn)]
         public void TestTFO()
         {
             this.RunPowerShellTest(
+                _logger,
                 Constants.NewModel,
                 "Test-TFO -vaultSettingsFilePath \"" +
                 this.vaultSettingsFilePath +
@@ -107,6 +114,7 @@ namespace RecoveryServices.SiteRecovery.Test
         public void TestPlannedFailover()
         {
             this.RunPowerShellTest(
+                _logger,
                 Constants.NewModel,
                 "Test-PlannedFailover -vaultSettingsFilePath \"" +
                 this.vaultSettingsFilePath +
@@ -120,6 +128,7 @@ namespace RecoveryServices.SiteRecovery.Test
         public void TestReprotect()
         {
             this.RunPowerShellTest(
+                _logger,
                 Constants.NewModel,
                 "Test-Reprotect -vaultSettingsFilePath \"" +
                 this.vaultSettingsFilePath +