+++ src/ResourceManager/Sql/Commands.Sql/Replication/Cmdlet/NewAzureSqlDatabaseCopy.cs
@@ -28,8 +28,8 @@ namespace Microsoft.Azure.Commands.Sql.Replication.Cmdlet
     /// <summary>
     /// Cmdlet to create a new Azure SQL Database Copy
     /// </summary>
-    [Cmdlet(VerbsCommon.New, "AzureRmSqlDatabaseCopy",
-        ConfirmImpact = ConfirmImpact.Low, SupportsShouldProcess = true, DefaultParameterSetName = DtuDatabaseParameterSet)]
+    [Cmdlet(VerbsCommon.New, "AzureRmSqlDatabaseCopy", ConfirmImpact = ConfirmImpact.Low, SupportsShouldProcess = true, 
+        DefaultParameterSetName = DtuDatabaseParameterSet), OutputType(typeof(AzureSqlDatabaseCopyModel))]
     public class NewAzureSqlDatabaseCopy : AzureSqlDatabaseCopyCmdletBase
     {
         private const string DtuDatabaseParameterSet = "DtuBasedDatabase";
@@ -103,7 +103,7 @@ namespace Microsoft.Azure.Commands.Sql.Replication.Cmdlet
         /// Gets or sets the compute generation of the database copy
         /// </summary>
         [Parameter(ParameterSetName = VcoreDatabaseParameterSet, Mandatory = true,
-            HelpMessage = "The compute generation of teh Azure Sql Database copy.")]
+            HelpMessage = "The compute generation to assign to the new copy.")]
         [Alias("Family")]
         [PSArgumentCompleter("Gen4", "Gen5")]
         [ValidateNotNullOrEmpty]